fix delete permission

+++ src/app/Http/Controllers/FrameController.php
@@ -134,6 +134,10 @@ class FrameController extends Controller
         if( ! $frame ) {
             return redirect()->back()->with('success',false)->with('message','Khung này không tồn tại.');
         }
+        if( $frame->user_id != auth()->id() ) {
+            abort(403);
+        }
+
         $frame->removeImages();
         $this->frame->delete($frame->id);
         return redirect()->back()->with('success',true)->with('message','Xóa thành công.');
